Project Manager Stuff
This contains the Weekly Project Schedule, Action Items, Weekly Reports and the Master Schedule for the course for all of us to review, and talk about. Please do not modify without talking to the entire group first and if modifying make a new version so we never loss the old stuff.